Create iter_1.py
This code completes the requirements of iteration one stated on the PLTW documentation. The code asks the user to choose between a chicken, beef, or tofu sandwich while giving the subsequent prices. If the user enters an invalid option, it will continually ask the user to try again until they input a valid choice (chicken, beef, or tofu). Finally, the program prints the users choice to make sure that there were no mistakes in the order.