Create iter_2.py
This is the second iteration of the program. I first instantiated a dictionary with each key being a menu item and each value being the price of the menu item at the top of the file. Then, I continued on the previous code by implementing a features that asked for whether someone wanted a beverage or not, and then asks for the preferred size if they did want one. At the end, the price of the order so far is printed by a for loop that goes through the prices dictionary at the beginning.