Create iter_3.py
This iteration adds the option for the user to order fries, and lets them choose the size of their fries as well. The program is not lenient with mistakes, and automatically adds large fries to the order if an incorrect order is made. If the user asks for small fries, the option to mega-size is given, and the program adds large fries to the order if the user asks to mega-size. The total cost is then added at the end.